refactored some auth components, backend

+++ apps/forum-backend/src/lib/queries/thread/remove-thread.ts
@@ -3,46 +3,67 @@ import { supabase } from "#shared/supabase/supabase-client.ts";
 import type { DB } from "@repo/types/db/forum-database-types.ts";
 import type { Transaction } from "kysely";
 
+async function deleteComments(id: string, trx: Transaction<DB>) {
+  const query = await trx
+    .deleteFrom("comments")
+    .where("parent_id", "=", id)
+    .where("parent_type", "=", "thread")
+    .executeTakeFirst()
+
+  if (!query.numDeletedRows) {
+    return false
+  }
+
+  return true;
+}
+
 async function removeThreadImages(id: string, trx: Transaction<DB>) {
+  let finished: boolean = false;
+
   const threadImages = await trx
     .selectFrom("threads_images")
     .select("image_url")
     .where("thread_id", "=", id)
     .execute();
 
-  if (threadImages) {
+  if (threadImages && threadImages.length >= 1) {
     threadImages.forEach(async (image) => {
       await supabase.storage.from("threads").remove([`/${image.image_url}`]);
     })
+
+    const result = await trx
+      .deleteFrom("threads_images")
+      .where("thread_id", "=", id)
+      .executeTakeFirstOrThrow();
+
+    if (result.numDeletedRows) {
+      finished = true
+    }
+
+    finished = false
+  } else {
+    finished = true;
   }
 
-  return await trx
-    .deleteFrom("threads_images")
-    .where("thread_id", "=", id)
-    .returning("id")
-    .executeTakeFirstOrThrow();
+  return finished
 }
 
 export const removeThread = async (id: string) => {
-  return await forumDB.transaction().execute(async (trx) => {
-    await removeThreadImages(id, trx);
+  const query = await forumDB.transaction().execute(async (trx) => {
+    const removeImages = await removeThreadImages(id, trx);
 
-    await trx
-      .deleteFrom('threads_users')
-      .where('thread_id', '=', id)
-      .returning("thread_id")
-      .executeTakeFirstOrThrow();
+    if (!removeImages) return;
 
-    await trx
-      .deleteFrom('threads_reactions')
-      .where('thread_id', '=', id)
-      .returning("id")
-      .executeTakeFirst();
+    const removeComments = await deleteComments(id, trx)
 
-    return await trx
+    if (!removeComments) return;
+
+    return trx
       .deleteFrom('threads')
       .where('id', '=', id)
       .returning("id")
       .executeTakeFirstOrThrow();
   });
+
+  return query;
 }
